huggingface-backend/src/requirements.txt
# FastAPI and Web Server
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database and ORM
sqlmodel>=0.0.14
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
aiosqlite>=0.19.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
bcrypt>=4.0.0

# Validation and Serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# CORS
fastapi-cors>=0.0.6

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
